% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage and Distance data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
\item{DISTANCE}{the distance from the fires station where this data was gathered, in miles}
\item{DAMAGE}{X 1000 dollars in damages caused by the fires}
...
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A set of data detailing the damage caused by fires and the distance from the fire station.
}
\keyword{datasets}
