% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinom.R
\name{mybin}
\alias{mybin}
\title{Binomial Function}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
A binomial and its barplot
}
\description{
Binomial Function
}
\examples{
mybin(iter = 500, n = 10, p = 0.7)
}
