% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{BootStrap Function}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of Iterations}

\item{x}{The given random sample}

\item{fun}{The mean}

\item{alpha}{The % Confidence Intervals}

\item{cx}{Text Label Size}

\item{...}{Extra Parameters}
}
\value{
A Histogram
}
\description{
BootStrap Function
}
\examples{
\dontrun{
sam <- rnorm(20, mean = 10, sd = 4)
myboot(iter = 10000, x = sam, fun = "mean", alpha = 0.05, cx = 1.5)
}
}
