% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Lower-Tail (P(X<=a)) Probablility Distrubution Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the distribution}

\item{sigma}{The standard deviation of the distribution}

\item{a}{The probablility of the samples being below this value}
}
\value{
A graph of the distribution with the probablility of a shaded in, as well its value.
}
\description{
Normal Lower-Tail (P(X<=a)) Probablility Distrubution Function
}
\examples{
myncurve(5, 10, 3)
}
