% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{myntickets}
\alias{myntickets}
\title{Solving an Overbooking Problem}
\usage{
myntickets(N = 200, gamma = 0.02, p = 0.95, ...)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{Probability that a plane will be overbooked}

\item{p}{Probability that a passenger will show}

\item{...}{Extra parameters if needed}
}
\value{
Two graphs that use discrete and continuous distributions, respectively.
}
\description{
Solving an Overbooking Problem
}
\examples{
myntickets(N = 400, gamma = 0.02, p = 0.95)
}
