% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{Quadratic Plot}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{Quadratic line equation}

\item{y}{Beta Values}
}
\value{
A quadratic plot
}
\description{
Quadratic Plot
}
\examples{
myplot(x = myQuadraticLineModel, y = 2)
}
